Updated project to clear up warnings, updated sample apps, and refactored some class names.

+++ source/Classes/UIEventTracking/ApigeeUIControlEvent.m
@@ -57,7 +57,7 @@
     
     if( control.tag > 0 )
     {
-        self.tag = control.tag;
+        self.tag = (int)control.tag;
     }
 }
